Small fixes for VM install

+++ scripts/vm-nixos-first-boot.sh
@@ -24,31 +24,49 @@ shift "$(($OPTIND - 1))"
 read -r
 
 nix --version
+echo $NIX_PATH
 
 # Fix command-not-found
 nix-channel --add https://nixos.org/channels/nixos-unstable nixos
 nix-channel --update
+printf "\rUpdated nix-channel \n"
+nix-channel --list
 
 # Fix command-not-found, but for root
 sudo nix-channel --add https://nixos.org/channels/nixos-unstable nixos
 sudo nix-channel --update
+printf "\rUpdated nix-channel \n"
+nix-channel --list
 
 # Fix missing home-manager
 nix-channel --add https://github.com/nix-community/home-manager/archive/master.tar.gz home-manager
 nix-channel --update
-nix-shell '<home-manager>' -A install # Note: you may need to logout and back in again
-
+printf "\rUpdated nix-channel \n"
 nix-channel --list
+
+hm=$(which home-manager)
+if [ -z $hm ]; then
+    printf "\rInstalling home-manager\n"
+    nix-shell '<home-manager>' -A install # Note: you may need to logout and back in again
+else
+    printf "\rhome-manager found already\n"
+fi
 
 # Look for ~/.nix-defexpr/channels
 echo $NIX_PATH
 
 # Avoid conflict. It shoud be empty
-rm -v /home/df/.config/fish/config.fish
+if [ -f "/home/df/.config/fish/config.fish" ]; then
+    printf "\rRemoving existing fish config\n"
+    rm -v /home/df/.config/fish/config.fish
+else
+    printf "\rNo existing fish config found\n"
+fi
 
 # Installing
 cd /home/$USERNAME/.dotfiles
 
+printf "\rhome-manager version: \n"
 home-manager --version
 home-manager switch --flake ".#$USERNAME@$TARGET_HOSTNAME"
